Fixed a (very sutil) bug in the hash slot calculation
It was being calculated in terms of the *tile position*
instead of the *tileChunk position*. This meant that two
tiles within the same tileChunk would get different hash slots
and therefore would instanciate different tilechunks.

This meant the tile hash slot has being absolutely filled with
tileChunks that has (on average) only one tile set!

This worked correctly for a few tiles but would have exploded
in a real scenario.