A little bit on API cleanup using const references
They're OK in the sense they document correctly
the usage of the argument and in general feel better.

It's strange to get the address of something in order to
pass a pointer when what I really want is a reference.

Passing pointer will still be used, but delegated to
when what I really want to pass is a pointer.
Otherwise, reference it is (it seems).