Created a hotEntity cache to refer to close entities
This is made to always have a small subset of entities
on which we iterate and do full update. This entities
are the ones that are closest to the player. The way that
subset is calculated is that when the player crosses
the tileChunk boundary, we iterate over all the "hot"
entities and unload them if they're outside a tileChunk
bound (a few tileChunks from the players tileChunk).

After that we search all the entities in order to make "hot"
the ones that actually are in the bounds.

TODO: Make a more intelligent spacial loop so we don't have
to check _ALL_ cold entities to determine which ones will
be hot now.